Updating IDs.

+++ commands/appeal.js
@@ -5,8 +5,8 @@ module.exports = {
     name: 'appeal',
     description: '[GENERAL] This command is restricted to the Appeals guild & channel(s) only. It will allow a member to appeal their ban punishment. <[setPrefix]appeal <appeal message>>',
     execute(message, args) {
-        //if (message.guild.id !== '781209099647451166') return;
-        //if (message.channel.id !== '781211090189025400') return;
+        if (message.guild.id !== '781209099647451166') return;
+        if (message.channel.id !== '781211090189025400') return;
 
         let receivedR = message.guild.roles.cache.find(role => role.name === "Appeal Received");
         if (message.member.roles.cache.has(receivedR.id)) return message.delete();
@@ -37,10 +37,10 @@ module.exports = {
             .setColor(`ff0000`)
             .setFooter(`ID: ${message.author.id}`)
 
-        message.guild.channels.cache.get(`781214229238775849`).send(appealEmbed).then(appealMSG => {
+        message.guild.channels.cache.get(`738863576890081340`).send(appealEmbed).then(appealMSG => {
             appealMSG.react('✅').then(appealMSG.react('🚫'));
 
-            message.member.roles.add('781209220329766943'); // UPDATE THIS WHEN IMPLEMENTED INTO GITHUB REPO
+            message.member.roles.add('691372147112673441');
             message.author.send(`**[${new Date().toLocaleTimeString()}] ${message.author.username}**, you have successfully appealed your ban for the **I Talk Server**. All appeals are reviewed and voted on Saturday, so please be patient.\nYour appeal message was submitted as the following:\n\`\`\`${appealMessage || 'No appeal message was sent.'}\`\`\``);
         });
 
@@ -67,7 +67,7 @@ module.exports = {
                         .setImage(appealAttachment.url)
                         .setColor(`c66523`)
 
-                    message.guild.channels.cache.get(`781214229238775849`).send(attachmentEmbed);
+                    message.guild.channels.cache.get(`738863576890081340`).send(attachmentEmbed);
                     message.author.send(`Attachment(s) provided with appeal:`, {
                         files: [appealAttachment.url]
                     });
